templates put

+++ backend/routers/templates.py
@@ -10,11 +10,7 @@ from backend.routers.exercises import get_exercise
 router = APIRouter()
 
 
-def validate_workout_template(cursor: Cursor, template: Workout_Template, username: str):
-    cursor.execute("SELECT * FROM template_workouts WHERE name = ? AND username = ?",(template.name, username))
-    if cursor.fetchone():
-        raise HTTPException(status_code=400, detail="Template name is currently in use")
-    
+def validate_workout_template(template: Workout_Template):
     if len(template.name) == 0:
         raise HTTPException(status_code=400, detail="Unnamed template")
     
@@ -69,6 +65,22 @@ def insert_template_set(cursor: Cursor, set_pos: int, set_template: Set_Template
     """, (exercise_template_id, set_template.reps, set_template.rep_range_start, set_template.rep_range_end, set_template.time_range_start, set_template.time_range_end, set_pos))
 
 
+def insert_template_exercises_and_sets(cursor: Cursor, template: Workout_Template):
+    for pos,exercise_template in enumerate(template.exercise_templates):
+        exercise_row: Row = get_exercise_row(cursor, exercise_template.exercise_id)
+        
+        exercise_template_id: int = insert_template_exercise(cursor, pos, exercise_template, template.id)
+        exercise_template.exercise_name = exercise_row["name"]
+
+        if not exercise_template.set_templates:
+            raise HTTPException(status_code=400, detail="Empty or null set templates array")
+        
+        for set_pos,set_template in enumerate(exercise_template.set_templates):
+            validate_set_template(set_template, exercise_row)
+                
+            insert_template_set(cursor, set_pos, set_template, exercise_template_id)
+
+
 @router.post("/templates/me/", response_model=Workout_Template, response_model_exclude_none=True)
 def create_user_template(
     template: Workout_Template,
@@ -76,26 +88,17 @@ def create_user_template(
     conn: Annotated[Connection, Depends(get_db)]
 ):
     cursor: Cursor = conn.cursor()
+    cursor.execute("SELECT * FROM template_workouts WHERE name = ? AND username = ?",(template.name, current_user.username))
+    if cursor.fetchone():
+        raise HTTPException(status_code=400, detail="Template name is currently in use")
 
-    validate_workout_template(cursor, template, current_user.username)
+    validate_workout_template(template)
     
     try:
         template.id = insert_template_workout(cursor, template, current_user.username)
         template.username = current_user.username
 
-        for pos,exercise_template in enumerate(template.exercise_templates):
-            exercise_row: Row = get_exercise_row(cursor, exercise_template.exercise_id)
-            
-            exercise_template_id: int = insert_template_exercise(cursor, pos, exercise_template, template.id)
-            exercise_template.exercise_name = exercise_row["name"]
-
-            if not exercise_template.set_templates:
-                raise HTTPException(status_code=400, detail="Empty or null set templates array")
-            
-            for set_pos,set_template in enumerate(exercise_template.set_templates):
-                validate_set_template(set_template, exercise_row)
-                    
-                insert_template_set(cursor, set_pos, set_template, exercise_template_id)
+        insert_template_exercises_and_sets(cursor, template)
 
         conn.commit()
     except HTTPException:
@@ -169,21 +172,6 @@ def convert_template_workouts_row_to_template(cursor: Cursor, template_workouts_
     )
 
 
-@router.get("/templates/me/{template_id}", response_model=Workout_Template, response_model_exclude_none=True)
-def get_user_template(
-    template_id: int,
-    current_user: Annotated[User, Depends(get_current_active_user)],
-    conn: Annotated[Connection, Depends(get_db)]
-):
-    cursor: Cursor = conn.cursor()
-    cursor.execute("SELECT * FROM template_workouts WHERE id = ? and username = ?", (template_id, current_user.username))
-    template_workouts_row: Row = cursor.fetchone()
-    if not template_workouts_row:
-        raise HTTPException(status_code=404, detail="Template not found")
-    
-    return convert_template_workouts_row_to_template(cursor, template_workouts_row)
-
-
 @router.get("/templates/me/", response_model=list[Workout_Template], response_model_exclude_none=True)
 def get_user_templates(
     current_user: Annotated[User, Depends(get_current_active_user)],
@@ -202,6 +190,61 @@ def get_user_templates(
     return templates
 
 
+@router.get("/templates/me/{template_id}", response_model=Workout_Template, response_model_exclude_none=True)
+def get_user_template(
+    template_id: int,
+    current_user: Annotated[User, Depends(get_current_active_user)],
+    conn: Annotated[Connection, Depends(get_db)]
+):
+    cursor: Cursor = conn.cursor()
+    cursor.execute("SELECT * FROM template_workouts WHERE id = ? and username = ?", (template_id, current_user.username))
+    template_workouts_row: Row = cursor.fetchone()
+    if not template_workouts_row:
+        raise HTTPException(status_code=404, detail="Template not found")
+    
+    return convert_template_workouts_row_to_template(cursor, template_workouts_row)
+
+
+@router.put("/templates/me/{template_id}", response_model=Workout_Template, response_model_exclude_none=True)
+def update_user_template(
+    template_id: int,
+    template: Workout_Template,
+    current_user: Annotated[User, Depends(get_current_active_user)],
+    conn: Annotated[Connection, Depends(get_db)]
+):
+    cursor: Cursor = conn.cursor()
+    cursor.execute("SELECT * FROM template_workouts WHERE id = ? AND username = ?", (template_id, current_user.username))
+    if not cursor.fetchone():
+        raise HTTPException(status_code=404, detail="Template not found")
+    
+    validate_workout_template(template)
+
+    try:
+        cursor.execute("""
+            UPDATE template_workouts
+            SET name = ?
+            WHERE id = ? AND username = ?
+        """,
+        (template.name, template_id, current_user.username))
+        template.id = template_id
+        template.username = current_user.username
+
+        cursor.execute("DELETE FROM template_exercises WHERE workout_template_id = ?", (template.id,))
+
+        insert_template_exercises_and_sets(cursor, template)
+
+        conn.commit()
+    except HTTPException:
+        conn.rollback()
+        raise
+    except Exception as e:
+        conn.rollback()
+        print(e)
+        raise HTTPException(status_code=500, detail="Internal server error")
+    
+    return template
+
+
 @router.delete("/templates/me/{template_id}", status_code=204)
 def delete_user_template(
     template_id: int,
@@ -209,10 +252,9 @@ def delete_user_template(
     conn: Annotated[Connection, Depends(get_db)]
 ):
     cursor: Cursor = conn.cursor()
-    
     cursor.execute("SELECT * FROM template_workouts WHERE id = ? AND username = ?", (template_id, current_user.username))
     if not cursor.fetchone():
-        raise HTTPException(status_code=404, detail="Workout template not found")
+        raise HTTPException(status_code=404, detail="Tempalte not found")
     
-    cursor.execute("DELETE FROM template_workouts WHERE id = ?", (template_id,))
+    cursor.execute("DELETE FROM template_workouts WHERE id = ? AND username = ?", (template_id, current_user.username))
     conn.commit()
